Add RMS envelope function for RPV calculation.
According to Power’s website
(https://www.jonathanpower.net/paper-multiecho.html), the RMS option
for MATLAB’s envelope function was used on the z-scored trace, with a
window of 4000 (equal to 10 seconds with 400Hz data). Then, the std of
the upper envelope was used as the RPV measure.

Quote from the site:
To calculate the shown envelopes in Matlab (for a 400Hz respiratory
trace):
```
% variability of respiratory envelope
z = zscore(respbelt);
[yu yl] = envelope(z,4000,'rms');
respvar = std(yu);
```